adsprpc: Handle UAF scenario in put_args
Currently, the DSP updates header buffers with unused DMA handle fds.
In the put_args section, if any DMA handle FDs are present in the
header buffer, the corresponding map is freed. However, since the
header buffer is exposed to users in unsigned PD, users can update
invalid FDs. If this invalid FD matches with any FD that is already
in use, it could lead to a use-after-free (UAF) vulnerability.
As a solution,add DMA handle references for DMA FDs, and the map for
the FD will be freed only when a reference is found.

Acked-by: Karthik Sanagavarapu <quic_kartsana@quicinc.com>
Change-Id: I86180d226afd771b414ea8da2bed409beb04a0e1
Signed-off-by: Ansa Ahmed <quic_ansa@quicinc.com>